Playing with swapNodes
I fixed some bugs, and played with swapNodes. we need to focus on the way changeX and changeY interacts with the ovals, sometimes they should be negative, sometimes positive. (it may depend on trickleUP and trickleDown.) 

For some reason parent nodes aren't moving, only child nodes (as in, cur moves but parents does not).

We also need to focus on where swapNodes is called in trickleUP and trickleDown and why its not being called multiple times when it should be (i.e. when a node needs to move several levels up, swapNode should be called for each parent that it surpasses so we can see that swap).