Clean up appearance of Array visualization
Changed how the position of the grid of cell values is calculated and drawn.

Added a separator line and beige background color to the operations area.

Fixed a bug in the animation of deleting a value where the first value slid 1 pixel too far to the left.   Standardized definition of fonts used.

Edited ToDo list to remove sorting related tasks and delete/insert at index