Added Notebook to show all visualizations in different tabbed panes
Strangely, importing packages in the same directory didn't work when
importing the entire PythonVisualizations library.  I had to import
them with a leading period to show they were relative.  That broke
loading them when importing while the current directory is
PythonVisualizations.  The result is, the code tries to load them
one way and fails over to the other method if it can't find them.