Added call stack visualization
This enhancement changes VisualizationApp's API for animated methods.  Each method creates a call environment for its local variables that are animated on the canvas.  Once the animation is done, it cleans up that call environment.  The VisualizationApp takes care of cleaning up any items left on the call stack including the verfy first call.