Enable argument help text for text entry widgets
This changes how a helpText label is created for the text entry widgets
and how it is removed by clicking on it or putting keyboard focus on
one of the text entry widgets.  The label is placed below all the text
entry widgets, and it only allows one label for all text entries.

A test program was added to the Test Results directory to see how the
layout of buttons progresses with each call to addOperation().