Recofigured bit vector display and probe arrows in BloomFliter
This makes BloomFilter a VisualizationApp subclass.  It displays the
bit vector on multiple rows using filled squares to indicate bits that
are set.  It draws arrows along spline curves to show the hash probes
going to the cells of the bit vector.  It maintains a list of inserted
words on the canvas.