Issue75-Complete
Shani and I implemented a array visualization for a min heap. We made two functions. "New" creates a new array of two cells. "Insert" adds a value to the end of the array. If it need to be swaped, it will be swapped and it will also grow the array if necessary in conjuction with the Heap.py code