Changed AlgorithmVisualizations to load modules in preferred order
Added a Loading ... message to the intro page to show progress in
loading modules.

Changed constructors of VisualizationApp subclasses to pass their
arguments for the superclass constructor in the kwargs dictionary.

Included OrderedArray, Stack, and SkipList in the PythonVisualizations
package.

Made SkipList constructor have a default maximum number of keys