Change constructor args for BinaryTreeBase to take bounding box
Made both endpoints of lines connecting tree nodes be at the center of
the node's circle and placed "behind" the root node's circle in the
stacking order.  Highlighted lines are placed "behind" the root node's
circle but in front of all the other lines.

Removed unused calculateParentLineCoordinates method.

Changed createHighlightedValue to accept an optional callEnviron parameter
like createHighlightedLine and createHighlightedCircle.  If the callEnviron
is passed, the new highlight item is added to it

Changed BinarySearchTreeBaseTest.py to exercise highlighted lines, circles,
and values, and include status messages.

Changed BinaryTree.py to use new constructor for BinaryTreeBase