Clean up heap structures after user stops animation
The heap customizes the cleanUp call to call fixPositions when the
call stack is empty (beginning of an operation).  The fixPositions method
restores any array cells and values that were moved and repositions the
nItems index

The insert method's grow stack algorithm was modified to only mark
the duplicates of the current cells and values as temporary for later
clean up.  This allows the lasting cells and values to be repositioned
by fixPositions if users stop the animations.