Fix delete on sorted list and text size and positions
This fixes the problem where deleting a non-existant key from a sorted
list would step through the entire list.  It now stops after finding
a key larger than the goal key.

The new textWidth and textHeight methods are used for determining
the size and position of text items, especiallly in traverse().  The
negative font sizes are used in specifications to get fonts sized in
pixels rather than points.