Handle Enter key in text entry arguments by running last operation
This commit adds a feature to VisualizationApp where it stores the
last operation run to be re-run when Return/Enter is typed in a text
entry widget.  To do this and be able to show the button being
pressed in the way it appears when there's a mouse click on it,
I had to introduce ttk.Buttons as the default button type for
operations.  On macOS, I couldn't find a way to temporarily make
a traditional Tk button show it activehighlight appearnce.

As a consequence, the way buttons are enabled and disabled changes.
The ttk.button and other widgets have a different API for managing
their state and appearance.  VisualizationApp now handles both and
changes to a widget's state should be made with the widgetState()
method.  The method allows getting and setting the state like the
other Tk methods.

A couple of apps had to be changed to use widgetState() (and one to
simply remove an unused method that invoked the Tk button interface).