Fix spindle appearance during moves and handle user stops better
These changes fix the problem where a shortened spindle would appear
floating above a stack of disks when one was being moved.

This also adds customized form of the stop button that ensures disks
are moved back to their spindle when a user stop occurs in the middle of
a move.  User stops can also happen when the labels are being moved
or the call stack is being changed (since there are calls to wait in
them).  That means that canvas items and the code text box can be
deleted during those waits, so I added more checks that the items
still exist before updating them.