Highlight bits that are	already	set during inserts
This change adds a temporary highlight to each bit in the vector that
was already set	 by a previous insertion.  The highlight  remains	after
the insertion completes	       and is erased by the next operation.

Also the contents of the internal __displayList	were simplified	to
hold the canvas	  item IDs for the bit vector rectangles.