Changed SortingBase and subclasses to use drawnValue
Made cell width adjust "continuously" based on number of cells
requested and current canvas width.  The width can be any integer
value between the minimum and maximum values defined in the class.
When the width gets narrower than the width needed to show the max
value, the values are hidden.  Fixed a couple of pointer movementss to
use cell width instead of cell size.

Updated some legacy code for mergesort in AdvancedSorting to use newer
constructs, but didn't test it.

Fixed a couple of hardcoded references to max value.