Reassignment now relies on SortingBase
When reassigning a number, ShellSort now relies on assignElement in SortingBase.py.

toBeRemoved is a list that is responsible for removing the item that is being copied over by assignElement.

SortingBase.py now has a parameter startAngle so it can move items on a curve if the angle is not 0