Add regular expression patterns and declaration name to code highlights
Instead of passing an exaxt string to highlightCode, this commit
enables passing a regular expression pattern (re.Pattern) to find
the code text to highlight.

When CodeHighlightBlock objects are initialized, it tries to find
a def or class definiition in the first line and stores the named
method, function, or class in the blockName attribute of the object.
That name will also show up in the string version of the object.
This is only useful for debugging.  Also for debugging purposes, the
getCodeHighlightBlock function is defined here.

The tests were expanded to
 - test regular expression patterns
 - print the fragment spans (and error messages) on the canvas
 - test the scrolling of the canvas by using canvasBounds
 - only handle expected exceptions and let others cause a stack trace