Fix visibleCanvas calculation and scrolling to see items
This fixes a bug in the visibleCanvas calculation that assumed the
left and top edge coordiantes were always zero.  The setCanvasBounds
method now helps ensure that the canvasBounds attribute says in synch
with the scrollregion of the canvas.

The moveItems* methods now can take a list of (non-moving) items to
keep in view while items are being moved.

The scrollSettingsToSee method now finds the maximal subset of items
that can be viewed when determining where to adjust scroll settings.