Set up resize handler for runAllVisualizationMenu.
When the main window is resized, the resize handler changes the
grid cell width and height for the selected visualization app
to follow the change.  Also always re-grid app windows with sticky
edges.

Also forced the menu button that shows the menu of visualization
apps to be hidden if a command line argument is used to select the
first app.

Tested only under macOS.