Fix scrollToSee's canvas size estimate & bounding box selection algorithm
The widgetDimensions method provides a slightly larger estimate of the
canvas' dimensions than getting the size from the visibleCanvas method.
That caused scrolling to not go far enough in some cases.

The algorithm for selecting a subset of bounding boxes that can be shown
within the visible canvas, was not thorough enough to explore all the
options needed.  Now it should find the maximum number of bboxes that
can be preserved.