Fixed findNearest(), added Traverse stub, and show insert vs update
The findNearest() program was always returning the nearPoint when
run without code showing (as in random fill operations).  This
caused it to occasionally fill points without paying attention to
the minimum separation.

The Traverse operation is stubbed in for later development.

The insert() operation now returns a flag indicating if a new
node was created or if an existing node was updated.  The status
message shows what happened.