Cut off	nodes moved beyond depth limit in binary trees
The BinaryTree.storeNodeTree routine now returns a tuple of all the
nodes that could not be stored in the tree array because they be
placed past the end of the array.  The Red-Black tree uses that
to rotate some nodes off the canvas and report their deletion.
Add a dashed line to the display to show cutoff level (hidden
unless window is enlarged).